working with using common in api; moved the rest of schemas to common; added index for common
