+++ api/db/apds.js
@@ -1,6 +1,7 @@
 const mongoose = require('mongoose');
 const { applyPatch } = require('fast-json-patch');
 const jsonpointer = require('jsonpointer');
+const { deepCopy } = require('@cms-eapd/common');
 const logger = require('../logger')('db/apds');
 const { updateStateProfile } = require('./states');
 const { validateApd } = require('../schemas');
@@ -42,7 +43,7 @@ const patchAPD = async (
   { APD = mongoose.model('APD') }
 ) => {
   // duplicate the apdDoc so that dates will be converted to strings
-  const apdJSON = JSON.parse(JSON.stringify(apdDoc));
+  const apdJSON = deepCopy(apdDoc);
   // apply the patches to the apd
   const { newDocument } = applyPatch(apdJSON, patch);
   // update the apd in the database
@@ -135,7 +136,7 @@ const updateAPDDocument = async (
 
     // Will probably eventually switch to apd.validate
     const validationErrors = {};
-    const valid = validate(JSON.parse(JSON.stringify(updatedDoc)));
+    const valid = validate(deepCopy(updatedDoc));
     if (!valid) {
       // Rather than send back the full error from the validator, pull out just the relevant bits
       // and fetch the value that's causing the error.
